refactor: rename to standard and portable recipes
- Rename modules: recipe_local → recipe, recipe_remote → recipe_portable
- Reorganize fixtures: flatten structure, rename remote → portable
- Update documentation: local → standard, remote → portable
- Add 9 portable recipe examples for acceptance testing
- Clarify conceptual model: portability (embedded runtime) vs standard (system runtimes)
- Update examples documentation to focus on caching use cases
