+++ docs/cache/recipes/portable/index.md
@@ -1,19 +1,25 @@
-# Remote Recipes
+# Portable Recipes
 
-Remote recipes allow you to execute portable JavaScript recipes directly from Git repositories without manually downloading or managing them. This enables easy sharing and reuse of build automation across teams and projects.
+Portable recipes are JavaScript files executed in Fabrik's embedded QuickJS runtime. They can be run locally or fetched directly from Git repositories, enabling easy sharing and reuse of build automation across teams and projects.
 
 > [!IMPORTANT]
-> **Why JavaScript?** Remote recipes use JavaScript because Fabrik embeds the **QuickJS runtime** (with AWS LLRT modules) directly into the binary. This means:
+> **Why JavaScript?** Portable recipes use JavaScript because Fabrik embeds the **QuickJS runtime** (with AWS LLRT modules) directly into the binary. This means:
 > - **Zero external dependencies** - No need to install Node.js, Deno, or any other runtime
 > - **Guaranteed cross-platform** - Works identically on macOS, Linux, and Windows
-> - **Fast startup** - QuickJS starts in milliseconds, making recipe execution instant
+> - **Fast startup** - QuickJS starts in milliseconds (~1ms vs ~50ms for Node.js)
 > - **Small binary size** - Embedded runtime adds minimal overhead to Fabrik's binary
 >
-> Unlike local script recipes (which require bash, python, etc. to be installed), remote recipes work out-of-the-box on any system with Fabrik installed.
+> Unlike standard recipes (which require bash, node, python, etc. to be installed), portable recipes work out-of-the-box on any system with Fabrik installed.
 
 ## Overview
 
-Remote recipes use the `@` prefix syntax to reference recipes stored in Git repositories:
+Portable recipes can be run locally:
+
+```bash
+fabrik run build.js
+```
+
+Or fetched from Git repositories using the `@` prefix syntax:
 
 ```bash
 fabrik run @org/repo/path/script.js
@@ -29,7 +35,7 @@ Fabrik automatically:
 
 ## Comparison with CI Reusable Steps
 
-Remote recipes share similarities with CI reusable steps (like GitHub Actions, GitLab CI Components, and Forgejo Actions) but are designed for a different purpose:
+Portable recipes share similarities with CI reusable steps (like GitHub Actions, GitLab CI Components, and Forgejo Actions) but are designed for a different purpose:
 
 **CI Reusable Steps** (GitHub Actions, GitLab CI Components, Forgejo Actions) are CI/CD workflows that:
 - Run in cloud infrastructure (provider-specific runners)
@@ -38,16 +44,16 @@ Remote recipes share similarities with CI reusable steps (like GitHub Actions, G
 - Execute in response to repository events (push, pull request, etc.)
 - Ideal for automated testing, deployment, and release workflows
 
-**Remote Recipes** are portable automation scripts that:
+**Portable Recipes** are portable automation scripts that:
 - Run locally on developer machines or in any CI environment
 - Not coupled to any specific CI provider
 - Use simple JavaScript with Fabrik's embedded runtime
 - Execute on-demand via `fabrik run` command
 - Ideal for cached build steps, code generation, and reproducible automation
 
-Think of remote recipes as **lightweight, portable actions** that work anywhere Fabrik is installed, with the added benefit of content-addressed caching for fast, incremental builds.
+Think of portable recipes as **lightweight, portable actions** that work anywhere Fabrik is installed, with the added benefit of content-addressed caching for fast, incremental builds.
 
-## Why Remote Recipes?
+## Why Portable Recipes?
 
 - **Easy Sharing** - Share recipes across teams by publishing them in Git repositories. No need to copy files manually.
 - **Version Control** - Pin recipes to specific versions using Git tags:
