+++ docs/cache/recipes/standard/examples.md
