+++ fixtures/cache/recipes/bash/build.sh
