+++ fixtures/cache/recipes/bash/deploy.sh
