+++ fixtures/cache/recipes/bash/output.txt
