+++ fixtures/cache/recipes/bash/simple.sh
