+++ fixtures/cache/recipes/bash/with-env.sh
