+++ fixtures/cache/recipes/bash/with-glob.sh
