+++ fixtures/cache/recipes/bash/with-inputs.sh
