+++ fixtures/cache/recipes/node/build.js
