+++ fixtures/cache/recipes/node/simple.js
