+++ fixtures/cache/recipes/node/with-env.js
