+++ fixtures/cache/recipes/node/with-inputs.js
