+++ fixtures/cache/recipes/portable/simple-echo.js
@@ -1,0 +1,5 @@
+// Simple recipe that just echoes a message
+// FABRIK cache ttl="1h"
+
+console.log("Hello from remote recipe!");
+console.log("This is a simple atomic test recipe");
