+++ fixtures/cache/recipes/portable/test-runner.js
@@ -1,0 +1,45 @@
+// Simulated test runner recipe
+// FABRIK input "src/**/*.{js,ts}"
+// FABRIK input "tests/**/*.test.{js,ts}"
+// FABRIK output "coverage/"
+// FABRIK env "CI"
+
+import { mkdirSync, writeFileSync } from 'fs';
+import { glob } from 'fabrik:fs';
+
+console.log("Running test suite...");
+
+// Find test files
+const testFiles = await glob("tests/**/*.test.{js,ts}");
+const sourceFiles = await glob("src/**/*.{js,ts}");
+
+console.log(`Found ${testFiles.length} test files`);
+console.log(`Found ${sourceFiles.length} source files`);
+
+// Simulate test execution
+const testResults = {
+  total: testFiles.length * 10, // Simulate 10 tests per file
+  passed: testFiles.length * 9,
+  failed: testFiles.length * 1,
+  duration: 2.5,
+  timestamp: new Date().toISOString(),
+  ci: process.env.CI === "true"
+};
+
+// Create coverage directory
+mkdirSync("coverage", { recursive: true });
+
+// Generate coverage report
+const coverage = {
+  lines: { total: sourceFiles.length * 100, covered: sourceFiles.length * 85, pct: 85 },
+  statements: { total: sourceFiles.length * 120, covered: sourceFiles.length * 102, pct: 85 },
+  functions: { total: sourceFiles.length * 20, covered: sourceFiles.length * 18, pct: 90 },
+  branches: { total: sourceFiles.length * 50, covered: sourceFiles.length * 40, pct: 80 }
+};
+
+writeFileSync("coverage/coverage-summary.json", JSON.stringify(coverage, null, 2));
+writeFileSync("coverage/test-results.json", JSON.stringify(testResults, null, 2));
+
+console.log(`Tests: ${testResults.passed}/${testResults.total} passed`);
+console.log(`Coverage: ${coverage.lines.pct}% lines, ${coverage.functions.pct}% functions`);
+console.log("Coverage report saved to coverage/");
