+++ src/lib.rs
@@ -10,8 +10,8 @@ pub mod config_discovery;
 pub mod config_expansion; // Environment variable expansion for config files
 pub mod logging;
 pub mod p2p; // P2P cache sharing
-pub mod recipe; // Portable recipes (QuickJS/JavaScript)
-pub mod script; // Script recipes (bash, python, etc.)
+pub mod recipe; // Script recipes with content-addressed caching (bash, node, python, etc.)
+pub mod recipe_portable; // Portable recipes executed in Fabrik's embedded JS runtime
 pub mod storage;
 pub mod xdg;
 
@@ -19,5 +19,5 @@ pub mod xdg;
 pub use auth::AuthProvider;
 pub use config::FabrikConfig;
 pub use config_discovery::{discover_config, hash_config, DaemonState};
-pub use recipe::RecipeExecutor;
+pub use recipe_portable::RecipeExecutor;
 pub use storage::{create_storage, default_cache_dir, FilesystemStorage, Storage};
