+++ src/recipe/cache.rs
@@ -1,390 +1,296 @@
-// Recipe cache APIs - Content-addressed caching for portable recipes
-//
-// This module provides runCached() and needsRun() APIs that recipes can use
-// for content-addressed caching of build operations.
-
-use anyhow::{Context, Result};
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-
-/// Configuration options for cache operations
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CacheOptions {
-    /// Input file patterns (globs) that affect cache key
-    #[serde(default)]
-    pub inputs: Vec<String>,
-
-    /// Output paths (files or directories) to cache/restore
-    #[serde(default)]
-    pub outputs: Vec<String>,
-
-    /// Environment variables that affect cache key
-    #[serde(default)]
-    pub env: Vec<String>,
-
-    /// Cache directory override (optional)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cache_dir: Option<String>,
-
-    /// Upstream cache servers override (optional)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub upstream: Option<Vec<String>>,
-
-    /// Cache TTL override (optional)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ttl: Option<String>,
-
-    /// Hash method: "content", "mtime", or "size"
-    #[serde(default = "default_hash_method")]
-    pub hash_method: String,
-}
-
-fn default_hash_method() -> String {
-    "content".to_string()
-}
-
-impl Default for CacheOptions {
-    fn default() -> Self {
-        Self {
-            inputs: Vec::new(),
-            outputs: Vec::new(),
-            env: Vec::new(),
-            cache_dir: None,
-            upstream: None,
-            ttl: None,
-            hash_method: default_hash_method(),
-        }
-    }
-}
-
-/// Result of a cache operation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[allow(dead_code)]
-pub struct CacheResult {
-    /// Was this a cache hit?
-    pub cached: bool,
-
-    /// The computed cache key (SHA256 hash)
-    pub cache_key: String,
-
-    /// Files that were restored (only for cache hits)
-    #[serde(default)]
-    pub restored_files: Vec<String>,
-
-    /// Duration in milliseconds (only for cache misses after action runs)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub duration_ms: Option<u64>,
-}
-
-/// Compute cache key from options
+/// Script cache storage and retrieval
 ///
-/// Cache key = SHA256(inputs_hash + env_values + hash_method)
-pub async fn compute_cache_key(options: &CacheOptions, working_dir: &Path) -> Result<String> {
-    let mut hasher = Sha256::new();
+/// Integrates with Fabrik's existing cache infrastructure to store/retrieve script outputs.
+use anyhow::{Context, Result};
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use std::fs;
+use std::path::{Path, PathBuf};
+use std::time::Duration;
 
-    // Hash input files
-    for pattern in &options.inputs {
-        let input_hash = hash_input_pattern(pattern, &options.hash_method, working_dir).await?;
-        hasher.update(input_hash.as_bytes());
-    }
+use super::outputs::ArchivedOutput;
 
-    // Hash environment variables
-    for env_var in &options.env {
-        if let Ok(value) = std::env::var(env_var) {
-            hasher.update(env_var.as_bytes());
-            hasher.update(b"=");
-            hasher.update(value.as_bytes());
-        }
-    }
-
-    // Include hash method in cache key
-    hasher.update(options.hash_method.as_bytes());
-
-    let hash = hasher.finalize();
-    Ok(hex::encode(hash))
+/// Cache entry metadata
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CacheMetadata {
+    pub version: u32,
+    pub cache_key: String,
+    pub script_path: String,
+    pub created_at: DateTime<Utc>,
+    pub expires_at: Option<DateTime<Utc>>,
+    pub execution: ExecutionInfo,
+    pub inputs: Vec<InputInfo>,
+    pub outputs: Vec<ArchivedOutput>,
+    pub environment: std::collections::HashMap<String, String>,
+    pub cache_info: CacheInfo,
 }
 
-/// Hash files matching a pattern
-async fn hash_input_pattern(
-    pattern: &str,
-    hash_method: &str,
-    working_dir: &Path,
-) -> Result<String> {
-    let mut hasher = Sha256::new();
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ExecutionInfo {
+    pub exit_code: i32,
+    pub duration_ms: u64,
+    pub runtime: String,
+    pub runtime_version: Option<String>,
+}
 
-    // Resolve pattern relative to working directory
-    let pattern_path = if Path::new(pattern).is_absolute() {
-        pattern.to_string()
-    } else {
-        working_dir.join(pattern).to_string_lossy().to_string()
-    };
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InputInfo {
+    pub glob: String,
+    pub files: Vec<String>,
+    pub combined_hash: String,
+}
 
-    // Find all matching files
-    let paths = glob::glob(&pattern_path)
-        .context("Failed to parse glob pattern")?
-        .filter_map(|entry| entry.ok())
-        .collect::<Vec<_>>();
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CacheInfo {
+    pub cache_hit: bool,
+    pub upstream_used: Option<String>,
+    pub restore_time_ms: Option<u64>,
+}
 
-    // Sort for deterministic hashing
-    let mut sorted_paths = paths;
-    sorted_paths.sort();
+/// Cache entry with metadata and archive path
+#[derive(Debug, Clone)]
+pub struct CacheEntry {
+    pub metadata: CacheMetadata,
+    pub archive_path: PathBuf,
+}
 
-    // Hash each file
-    for path in sorted_paths {
-        if !path.is_file() {
-            continue;
+/// Script cache manager
+pub struct ScriptCache {
+    cache_dir: PathBuf,
+}
+
+impl ScriptCache {
+    /// Create new script cache manager
+    pub fn new(cache_dir: PathBuf) -> Result<Self> {
+        let script_cache_dir = cache_dir.join("scripts");
+        fs::create_dir_all(&script_cache_dir).with_context(|| {
+            format!(
+                "Failed to create cache directory: {}",
+                script_cache_dir.display()
+            )
+        })?;
+
+        Ok(Self {
+            cache_dir: script_cache_dir,
+        })
+    }
+
+    /// Get cache entry if it exists and is not expired
+    pub fn get(&self, cache_key: &str) -> Result<Option<CacheEntry>> {
+        let entry_dir = self.cache_dir.join(cache_key);
+
+        if !entry_dir.exists() {
+            return Ok(None);
         }
 
-        // Include relative path in hash for uniqueness
-        let rel_path = path
-            .strip_prefix(working_dir)
-            .unwrap_or(&path)
-            .to_string_lossy();
-        hasher.update(rel_path.as_bytes());
+        let metadata_path = entry_dir.join("metadata.json");
+        let archive_path = entry_dir.join("outputs.tar.zst");
 
-        match hash_method {
-            "content" => {
-                // Hash file content
-                let content = tokio::fs::read(&path)
-                    .await
-                    .context("Failed to read input file")?;
-                hasher.update(&content);
+        if !metadata_path.exists() || !archive_path.exists() {
+            // Invalid cache entry - clean it up
+            let _ = fs::remove_dir_all(&entry_dir);
+            return Ok(None);
+        }
+
+        // Load metadata
+        let metadata_json = fs::read_to_string(&metadata_path)
+            .with_context(|| format!("Failed to read metadata: {}", metadata_path.display()))?;
+
+        let metadata: CacheMetadata =
+            serde_json::from_str(&metadata_json).context("Failed to parse metadata JSON")?;
+
+        // Check if expired
+        if let Some(expires_at) = metadata.expires_at {
+            if Utc::now() > expires_at {
+                // Expired - remove and return None
+                let _ = fs::remove_dir_all(&entry_dir);
+                return Ok(None);
             }
-            "mtime" => {
-                // Hash modification time
-                let metadata = tokio::fs::metadata(&path)
-                    .await
-                    .context("Failed to read file metadata")?;
-                if let Ok(mtime) = metadata.modified() {
-                    if let Ok(duration) = mtime.duration_since(std::time::UNIX_EPOCH) {
-                        hasher.update(duration.as_secs().to_le_bytes());
+        }
+
+        Ok(Some(CacheEntry {
+            metadata,
+            archive_path,
+        }))
+    }
+
+    /// Store cache entry
+    pub fn put(&self, cache_key: &str, metadata: CacheMetadata, archive_path: &Path) -> Result<()> {
+        let entry_dir = self.cache_dir.join(cache_key);
+        fs::create_dir_all(&entry_dir).with_context(|| {
+            format!("Failed to create entry directory: {}", entry_dir.display())
+        })?;
+
+        // Write metadata
+        let metadata_path = entry_dir.join("metadata.json");
+        let metadata_json =
+            serde_json::to_string_pretty(&metadata).context("Failed to serialize metadata")?;
+        fs::write(&metadata_path, metadata_json)
+            .with_context(|| format!("Failed to write metadata: {}", metadata_path.display()))?;
+
+        // Copy archive
+        let dest_archive = entry_dir.join("outputs.tar.zst");
+        fs::copy(archive_path, &dest_archive)
+            .with_context(|| format!("Failed to copy archive to: {}", dest_archive.display()))?;
+
+        Ok(())
+    }
+
+    /// Remove cache entry
+    pub fn remove(&self, cache_key: &str) -> Result<()> {
+        let entry_dir = self.cache_dir.join(cache_key);
+
+        if entry_dir.exists() {
+            fs::remove_dir_all(&entry_dir).with_context(|| {
+                format!("Failed to remove cache entry: {}", entry_dir.display())
+            })?;
+        }
+
+        Ok(())
+    }
+
+    /// List all cache entries
+    pub fn list(&self) -> Result<Vec<String>> {
+        let mut entries = Vec::new();
+
+        if !self.cache_dir.exists() {
+            return Ok(entries);
+        }
+
+        for entry in fs::read_dir(&self.cache_dir).with_context(|| {
+            format!(
+                "Failed to read cache directory: {}",
+                self.cache_dir.display()
+            )
+        })? {
+            let entry = entry?;
+            if entry.file_type()?.is_dir() {
+                if let Some(name) = entry.file_name().to_str() {
+                    entries.push(name.to_string());
+                }
+            }
+        }
+
+        entries.sort();
+        Ok(entries)
+    }
+
+    /// Get cache statistics
+    pub fn stats(&self) -> Result<CacheStats> {
+        let mut total_entries = 0;
+        let mut total_size = 0;
+        let mut total_files = 0;
+
+        if !self.cache_dir.exists() {
+            return Ok(CacheStats {
+                total_entries,
+                total_size_bytes: total_size,
+                total_files,
+            });
+        }
+
+        for entry in fs::read_dir(&self.cache_dir)? {
+            let entry = entry?;
+            if entry.file_type()?.is_dir() {
+                total_entries += 1;
+
+                // Read metadata
+                let metadata_path = entry.path().join("metadata.json");
+                if let Ok(metadata_json) = fs::read_to_string(&metadata_path) {
+                    if let Ok(metadata) = serde_json::from_str::<CacheMetadata>(&metadata_json) {
+                        for output in &metadata.outputs {
+                            total_size += output.size_bytes;
+                            total_files += output.file_count;
+                        }
                     }
                 }
             }
-            "size" => {
-                // Hash file size
-                let metadata = tokio::fs::metadata(&path)
-                    .await
-                    .context("Failed to read file metadata")?;
-                hasher.update(metadata.len().to_le_bytes());
-            }
-            _ => {
-                anyhow::bail!("Unknown hash method: {}", hash_method);
-            }
-        }
-    }
-
-    let hash = hasher.finalize();
-    Ok(hex::encode(hash))
-}
-
-/// KV store for tracking cache keys
-pub struct KvStore {
-    store_path: PathBuf,
-}
-
-impl KvStore {
-    /// Create a new KV store
-    pub fn new(cache_dir: &Path) -> Self {
-        let store_path = cache_dir.join("kv.json");
-        Self { store_path }
-    }
-
-    /// Load KV store from disk
-    async fn load(&self) -> Result<HashMap<String, serde_json::Value>> {
-        if !self.store_path.exists() {
-            return Ok(HashMap::new());
         }
 
-        let content = tokio::fs::read_to_string(&self.store_path)
-            .await
-            .context("Failed to read KV store")?;
-        let map: HashMap<String, serde_json::Value> =
-            serde_json::from_str(&content).context("Failed to parse KV store")?;
-        Ok(map)
+        Ok(CacheStats {
+            total_entries,
+            total_size_bytes: total_size,
+            total_files,
+        })
     }
 
-    /// Save KV store to disk
-    async fn save(&self, map: &HashMap<String, serde_json::Value>) -> Result<()> {
-        // Ensure parent directory exists
-        if let Some(parent) = self.store_path.parent() {
-            tokio::fs::create_dir_all(parent)
-                .await
-                .context("Failed to create KV store directory")?;
+    /// Clean all cache entries
+    #[allow(dead_code)]
+    pub fn clean_all(&self) -> Result<()> {
+        if self.cache_dir.exists() {
+            fs::remove_dir_all(&self.cache_dir).with_context(|| {
+                format!(
+                    "Failed to remove cache directory: {}",
+                    self.cache_dir.display()
+                )
+            })?;
+            fs::create_dir_all(&self.cache_dir).with_context(|| {
+                format!(
+                    "Failed to recreate cache directory: {}",
+                    self.cache_dir.display()
+                )
+            })?;
         }
-
-        let content = serde_json::to_string_pretty(map).context("Failed to serialize KV store")?;
-        tokio::fs::write(&self.store_path, content)
-            .await
-            .context("Failed to write KV store")?;
-        Ok(())
-    }
-
-    /// Check if key exists
-    pub async fn has(&self, key: &str) -> Result<bool> {
-        let map = self.load().await?;
-        Ok(map.contains_key(key))
-    }
-
-    /// Get value for key
-    pub async fn get(&self, key: &str) -> Result<Option<serde_json::Value>> {
-        let map = self.load().await?;
-        Ok(map.get(key).cloned())
-    }
-
-    /// Set value for key
-    pub async fn set(&self, key: &str, value: serde_json::Value) -> Result<()> {
-        let mut map = self.load().await?;
-        map.insert(key.to_string(), value);
-        self.save(&map).await?;
         Ok(())
     }
 }
 
-/// Check if action needs to run (uses only KV storage)
-///
-/// Returns true if cache miss (action should run)
-/// Returns false if cache hit (action can be skipped)
-pub async fn needs_run(options: CacheOptions, working_dir: &Path) -> Result<bool> {
-    // Compute cache key
-    let cache_key = compute_cache_key(&options, working_dir).await?;
-
-    // Determine cache directory
-    let cache_dir = if let Some(ref dir) = options.cache_dir {
-        PathBuf::from(dir)
-    } else {
-        working_dir.join(".fabrik/cache")
-    };
-
-    // Check KV store
-    let kv = KvStore::new(&cache_dir);
-    let exists = kv.has(&cache_key).await?;
-
-    // Return true if needs run (cache miss)
-    Ok(!exists)
+/// Cache statistics
+#[derive(Debug, Clone)]
+pub struct CacheStats {
+    pub total_entries: usize,
+    pub total_size_bytes: u64,
+    pub total_files: usize,
 }
 
-/// Archive outputs to cache
-pub async fn archive_outputs(
-    outputs: &[String],
-    cache_dir: &Path,
-    cache_key: &str,
-    working_dir: &Path,
-) -> Result<Vec<String>> {
-    let mut archived = Vec::new();
-
-    // Create archive directory
-    let archive_dir = cache_dir.join("artifacts").join(cache_key);
-    tokio::fs::create_dir_all(&archive_dir)
-        .await
-        .context("Failed to create archive directory")?;
-
-    for output_pattern in outputs {
-        let output_path = working_dir.join(output_pattern);
-
-        if output_path.is_file() {
-            // Archive single file
-            let dest = archive_dir.join(
-                output_path
-                    .file_name()
-                    .context("Invalid output file name")?,
-            );
-            tokio::fs::copy(&output_path, &dest)
-                .await
-                .context("Failed to archive output file")?;
-            archived.push(output_pattern.clone());
-        } else if output_path.is_dir() {
-            // Archive directory recursively
-            copy_dir_all(&output_path, &archive_dir.join(output_pattern))
-                .await
-                .context("Failed to archive output directory")?;
-            archived.push(output_pattern.clone());
-        }
-    }
-
-    Ok(archived)
+/// Parameters for creating cache metadata
+#[allow(clippy::too_many_arguments)]
+pub struct CreateMetadataParams<'a> {
+    pub cache_key: String,
+    pub script_path: &'a Path,
+    pub exit_code: i32,
+    pub duration: Duration,
+    pub runtime: String,
+    pub runtime_version: Option<String>,
+    pub outputs: Vec<ArchivedOutput>,
+    pub env_vars: &'a [String],
+    pub ttl: Option<Duration>,
 }
 
-/// Restore outputs from cache
-pub async fn restore_outputs(
-    outputs: &[String],
-    cache_dir: &Path,
-    cache_key: &str,
-    working_dir: &Path,
-) -> Result<Vec<String>> {
-    let mut restored = Vec::new();
-    let archive_dir = cache_dir.join("artifacts").join(cache_key);
+/// Helper to create cache metadata
+pub fn create_metadata(params: CreateMetadataParams) -> CacheMetadata {
+    let created_at = Utc::now();
+    let expires_at = params
+        .ttl
+        .map(|ttl| created_at + chrono::Duration::from_std(ttl).unwrap());
 
-    if !archive_dir.exists() {
-        return Ok(restored);
+    let environment = params
+        .env_vars
+        .iter()
+        .filter_map(|var| std::env::var(var).ok().map(|value| (var.clone(), value)))
+        .collect();
+
+    CacheMetadata {
+        version: 1,
+        cache_key: params.cache_key,
+        script_path: params.script_path.display().to_string(),
+        created_at,
+        expires_at,
+        execution: ExecutionInfo {
+            exit_code: params.exit_code,
+            duration_ms: params.duration.as_millis() as u64,
+            runtime: params.runtime,
+            runtime_version: params.runtime_version,
+        },
+        inputs: Vec::new(), // TODO: populate from input hashes
+        outputs: params.outputs,
+        environment,
+        cache_info: CacheInfo {
+            cache_hit: false,
+            upstream_used: None,
+            restore_time_ms: None,
+        },
     }
-
-    for output_pattern in outputs {
-        let archived_path = archive_dir.join(output_pattern);
-        let dest_path = working_dir.join(output_pattern);
-
-        if archived_path.is_file() {
-            // Restore single file
-            if let Some(parent) = dest_path.parent() {
-                tokio::fs::create_dir_all(parent)
-                    .await
-                    .context("Failed to create output directory")?;
-            }
-            tokio::fs::copy(&archived_path, &dest_path)
-                .await
-                .context("Failed to restore output file")?;
-            restored.push(output_pattern.clone());
-        } else if archived_path.is_dir() {
-            // Restore directory recursively
-            copy_dir_all(&archived_path, &dest_path)
-                .await
-                .context("Failed to restore output directory")?;
-            restored.push(output_pattern.clone());
-        }
-    }
-
-    Ok(restored)
-}
-
-/// Copy directory recursively
-fn copy_dir_all<'a>(
-    src: &'a Path,
-    dst: &'a Path,
-) -> std::pin::Pin<Box<dyn std::future::Future<Output = Result<()>> + Send + 'a>> {
-    Box::pin(async move {
-        tokio::fs::create_dir_all(dst)
-            .await
-            .context("Failed to create destination directory")?;
-
-        let mut entries = tokio::fs::read_dir(src)
-            .await
-            .context("Failed to read source directory")?;
-
-        while let Some(entry) = entries
-            .next_entry()
-            .await
-            .context("Failed to read directory entry")?
-        {
-            let file_type = entry
-                .file_type()
-                .await
-                .context("Failed to read file type")?;
-            let src_path = entry.path();
-            let dst_path = dst.join(entry.file_name());
-
-            if file_type.is_dir() {
-                copy_dir_all(&src_path, &dst_path).await?;
-            } else {
-                tokio::fs::copy(&src_path, &dst_path)
-                    .await
-                    .context("Failed to copy file")?;
-            }
-        }
-
-        Ok(())
-    })
 }
 
 #[cfg(test)]
@@ -392,123 +298,105 @@ mod tests {
     use super::*;
     use tempfile::TempDir;
 
-    #[tokio::test]
-    async fn test_compute_cache_key() {
-        let temp_dir = TempDir::new().unwrap();
-        let test_file = temp_dir.path().join("input.txt");
-        tokio::fs::write(&test_file, b"test content").await.unwrap();
+    #[test]
+    fn test_cache_put_and_get() {
+        let temp = TempDir::new().unwrap();
+        let cache_dir = temp.path().join("cache");
 
-        let options = CacheOptions {
-            inputs: vec!["input.txt".to_string()],
-            env: vec!["TEST_VAR".to_string()],
-            ..Default::default()
-        };
+        let cache = ScriptCache::new(cache_dir).unwrap();
 
-        std::env::set_var("TEST_VAR", "test_value");
+        // Create test archive
+        let archive_path = temp.path().join("test.tar.zst");
+        fs::write(&archive_path, b"test archive data").unwrap();
 
-        let key1 = compute_cache_key(&options, temp_dir.path()).await.unwrap();
-        let key2 = compute_cache_key(&options, temp_dir.path()).await.unwrap();
+        // Create metadata
+        let metadata = create_metadata(CreateMetadataParams {
+            cache_key: "test-key-123".to_string(),
+            script_path: Path::new("/path/to/script.sh"),
+            exit_code: 0,
+            duration: Duration::from_secs(10),
+            runtime: "bash".to_string(),
+            runtime_version: None,
+            outputs: Vec::new(),
+            env_vars: &[],
+            ttl: None,
+        });
 
-        // Same inputs should produce same key
-        assert_eq!(key1, key2);
-
-        // Different env var should produce different key
-        std::env::set_var("TEST_VAR", "different_value");
-        let key3 = compute_cache_key(&options, temp_dir.path()).await.unwrap();
-        assert_ne!(key1, key3);
-    }
-
-    #[tokio::test]
-    async fn test_kv_store() {
-        let temp_dir = TempDir::new().unwrap();
-        let kv = KvStore::new(temp_dir.path());
-
-        // Initially empty
-        assert!(!kv.has("test_key").await.unwrap());
-
-        // Set value
-        kv.set("test_key", serde_json::json!({"foo": "bar"}))
-            .await
+        // Store
+        cache
+            .put("test-key-123", metadata.clone(), &archive_path)
             .unwrap();
 
-        // Check exists
-        assert!(kv.has("test_key").await.unwrap());
+        // Retrieve
+        let entry = cache.get("test-key-123").unwrap();
+        assert!(entry.is_some());
 
-        // Get value
-        let value = kv.get("test_key").await.unwrap().unwrap();
-        assert_eq!(value, serde_json::json!({"foo": "bar"}));
+        let entry = entry.unwrap();
+        assert_eq!(entry.metadata.cache_key, "test-key-123");
+        assert!(entry.archive_path.exists());
     }
 
-    #[tokio::test]
-    async fn test_needs_run() {
-        let temp_dir = TempDir::new().unwrap();
-        let test_file = temp_dir.path().join("input.txt");
-        tokio::fs::write(&test_file, b"test content").await.unwrap();
+    #[test]
+    fn test_cache_expiry() {
+        let temp = TempDir::new().unwrap();
+        let cache_dir = temp.path().join("cache");
 
-        let options = CacheOptions {
-            inputs: vec!["input.txt".to_string()],
-            cache_dir: Some(temp_dir.path().join("cache").to_string_lossy().to_string()),
-            ..Default::default()
-        };
+        let cache = ScriptCache::new(cache_dir).unwrap();
 
-        // First run - should need to run
-        assert!(needs_run(options.clone(), temp_dir.path()).await.unwrap());
+        let archive_path = temp.path().join("test.tar.zst");
+        fs::write(&archive_path, b"test archive data").unwrap();
 
-        // Simulate cache hit by setting KV entry
-        let cache_key = compute_cache_key(&options, temp_dir.path()).await.unwrap();
-        let cache_dir = temp_dir.path().join("cache");
-        let kv = KvStore::new(&cache_dir);
-        kv.set(&cache_key, serde_json::json!({"timestamp": 123456}))
-            .await
-            .unwrap();
+        // Create metadata with immediate expiry
+        let mut metadata = create_metadata(CreateMetadataParams {
+            cache_key: "test-key-123".to_string(),
+            script_path: Path::new("/path/to/script.sh"),
+            exit_code: 0,
+            duration: Duration::from_secs(10),
+            runtime: "bash".to_string(),
+            runtime_version: None,
+            outputs: Vec::new(),
+            env_vars: &[],
+            ttl: Some(Duration::from_secs(0)),
+        });
 
-        // Second run - should not need to run
-        assert!(!needs_run(options, temp_dir.path()).await.unwrap());
+        // Force expiry in the past
+        metadata.expires_at = Some(Utc::now() - chrono::Duration::seconds(1));
+
+        cache.put("test-key-123", metadata, &archive_path).unwrap();
+
+        // Should return None (expired)
+        let entry = cache.get("test-key-123").unwrap();
+        assert!(entry.is_none());
     }
 
-    #[tokio::test]
-    async fn test_archive_and_restore() {
-        let temp_dir = TempDir::new().unwrap();
-        let working_dir = temp_dir.path().join("work");
-        let cache_dir = temp_dir.path().join("cache");
-        tokio::fs::create_dir_all(&working_dir).await.unwrap();
+    #[test]
+    fn test_cache_list() {
+        let temp = TempDir::new().unwrap();
+        let cache_dir = temp.path().join("cache");
 
-        // Create output file
-        let output_file = working_dir.join("output.txt");
-        tokio::fs::write(&output_file, b"build output")
-            .await
-            .unwrap();
+        let cache = ScriptCache::new(cache_dir).unwrap();
 
-        // Archive
-        let archived = archive_outputs(
-            &["output.txt".to_string()],
-            &cache_dir,
-            "test_key",
-            &working_dir,
-        )
-        .await
-        .unwrap();
+        let archive_path = temp.path().join("test.tar.zst");
+        fs::write(&archive_path, b"test archive data").unwrap();
 
-        assert_eq!(archived.len(), 1);
+        // Store multiple entries
+        for i in 1..=3 {
+            let key = format!("test-key-{}", i);
+            let metadata = create_metadata(CreateMetadataParams {
+                cache_key: key.clone(),
+                script_path: Path::new("/path/to/script.sh"),
+                exit_code: 0,
+                duration: Duration::from_secs(10),
+                runtime: "bash".to_string(),
+                runtime_version: None,
+                outputs: Vec::new(),
+                env_vars: &[],
+                ttl: None,
+            });
+            cache.put(&key, metadata, &archive_path).unwrap();
+        }
 
-        // Remove original
-        tokio::fs::remove_file(&output_file).await.unwrap();
-        assert!(!output_file.exists());
-
-        // Restore
-        let restored = restore_outputs(
-            &["output.txt".to_string()],
-            &cache_dir,
-            "test_key",
-            &working_dir,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(restored.len(), 1);
-        assert!(output_file.exists());
-
-        let content = tokio::fs::read_to_string(&output_file).await.unwrap();
-        assert_eq!(content, "build output");
+        let entries = cache.list().unwrap();
+        assert_eq!(entries.len(), 3);
     }
 }
