+++ src/recipe_portable/mod.rs
@@ -1,0 +1,12 @@
+// Portable recipes - Cross-platform JavaScript automation
+//
+// This module provides cross-platform build automation using JavaScript that runs
+// on Fabrik's embedded QuickJS runtime with custom Fabrik APIs exposed.
+
+pub mod cache;
+pub mod executor;
+pub mod remote;
+pub mod runtime;
+
+pub use executor::RecipeExecutor;
+pub use remote::RemoteRecipe;
